transport: Move valve from before the decoder to afterwards
If we have the valve before the decoder we will drop all
encoded frames until the stream is linked, including the
initial keyframe usually. So we have to request a new
keyframe which might or might not work but in any case
will take some extra time and delay the initial setup
of the pipeline.