transport: Always set up RTCP mux and normal RTCP pipelines and switch dynamically
Previously we were only setting up the non-RTCP mux pipeline unless
rtcp-mux was TRUE from the very beginning. This will never be the case
with WebRTC though, as we can only know if the other side can do RTCP mux
after SDPs are exchanged.

Now we always set up both paths, and for the sender dynamically switch
between RTCP muxing and normal RTCP depending on the property of the
media session. For receiving we just receive from both.