transport: Add a capsfilter after the encoder
And repurpose the now unused _owr_payload_create_encoded_caps() to
get the caps for it.

Here we now tell the encoder to create streams with a specific
profile or level. Currently this is only done for H264 to enforce
baseline profile, as anything else might not be supported by decoders.