OwrWindowRegistry: Implement registry for mapping tag to window handle
This is just a hash table wrapper with API. The OpenWebRTC API user will
be expected to register tags against window handles before renderers are
created. When the renderer is created, a tag is passed to its
constructor. This tag can then be used to look up the window handle the
renderer is supposed to render into.