transport: Don't put a capsfilter after the audio decoder and link the depayloader
We can't just force the audio decoder to output something specific, this will
fail with most audio decoders out there... and if conversion is required we
will do that later already.

Also make sure the pad that is ghosted is the decoded_valve and not the
capsfilter, which is incorrect.