transport: Be a bit smarter about setting up rtpbin for sending
This makes us not indiscriminately set up the rtpbin as a sender, but
only when we require it. We do need to make sure that we set up RTCP
sending even if we're not sending RTP, though.

Other than just cleaning up, this is also needed because rtpbin does not
send out RTCP packets when enabled as a sender, until the first RTP
packet is sent. In the case where we only receive media, this means RTCP
packets would never be sent out.