local: Scale on local media source if required
This allows us to support resolutions that are not supported by the
local source, by adding a videoscale if required after the source
itself.

This does mean we now have 2 'videoscale' elements -- one as part of the
OwrLocalMediaSource, and one from OwrMediaSource. This should remain as
it doesn't have a significant impact, and the second scaler will
probably be needed in the future when we support requests for multiple
resolutions from a single source.