media source: optional OpenGL support
Enable OpenGL for peers that support it. This means that OWR will now use
GLMemory between source -> renderer and source -> apple h264 encoders.

The implication is that in the above cases color conversion and scaling are done
via GL shaders. Also on Mac and iOS, memory is shared directly between the
camera and the renderer/encoders saving expensive GPU => sysmem memory copies.