owr_video_renderer: GL context request support
When the application using the renderer has a GL video sink using a
shared GL context and display, the NEED_CONTEXT synchronous message
emitted on the renderer pipeline bus has to be relayed to the
application.

Using a GClosure, OpenWebRTC can request the needed context from the
application side, which has to register a GCallback that returns a
pointer to a GstContext structure.