transport_agent: new API to explicitely start the pipeline
The owr_transport_agent_add_session function is now limited to storing
the session in an internal list. The list is processed by a new
owr_transport_agent_start function which will perform the necessary
operations in one go to prepare the GStreamer pipeline and start
playback.

This change of behavior is needed for the RTP bundling support because
the agent needs to be aware of all the sessions it will manage as soon
as rtpbin starts prerolling.